adding StaticL2Test.php
