+++ tests/LCacheTest.php
@@ -36,14 +36,7 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $this->dbh->exec('CREATE INDEX ' . $prefix . 'rewritten_entry ON ' . $prefix . 'lcache_tags ("event_id")');
     }
 
-    public function testClearStaticL2()
-    {
-        $l2 = new StaticL2();
-        $myaddr = new Address('mybin', 'mykey');
-        $l2->set('mypool', $myaddr, 'myvalue');
-        $l2->delete('mypool', new Address());
-        $this->assertNull($l2->get($myaddr));
-    }
+
 
     public function testStaticL2Expiration()
     {
@@ -250,6 +243,17 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $this->performCaughtUnserializationOnGetTest($l2);
     }
 
+    // Callers should expect an UnserializationException.
+    protected function performFailedUnserializationOnGetTest($l2)
+    {
+        $l1 = new StaticL1();
+        $pool = new Integrated($l1, $l2);
+        $invalid_object = 'O:10:"HelloWorl":0:{}';
+        $myaddr = new Address('mybin', 'performFailedUnserializationOnGetTest');
+        $l2->set('anypool', $myaddr, $invalid_object, null, [], true);
+        $pool->get($myaddr);
+    }
+
     /**
      * @expectedException LCache\UnserializationException
      */
@@ -267,6 +271,17 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
     {
         $l2 = new StaticL2();
         $this->performFailedUnserializationOnGetTest($l2);
+    }
+
+    public function performGarbageCollectionTest($l2)
+    {
+        $pool = new Integrated(new StaticL1(), $l2);
+        $myaddr = new Address('mybin', 'mykey');
+        $this->assertEquals(0, $l2->countGarbage());
+        $pool->set($myaddr, 'myvalue', -1);
+        $this->assertEquals(1, $l2->countGarbage());
+        $pool->collectGarbage();
+        $this->assertEquals(0, $l2->countGarbage());
     }
 
     public function testDatabaseL2GarbageCollection()
